pyproject.toml
[tool.poetry]
name = "robotframework-libtoc"
version = "1.3.0"
description = "Docs and TOC generator for Robot Framework resources and libs"
authors = ["Andre Mochinin"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/amochin/robotframework-libtoc"
include = ["robotframework_libtoc/*template.html"]

[tool.poetry.dependencies]
python = ">=3.7"
robotframework = ">=4"
importlib-resources = "~5.12"

[tool.poetry.scripts]
libtoc = "robotframework_libtoc.libtoc:main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
